requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2

sklearn-crfsuite==0.3.6
python-crfsuite==0.9.9

torch 
pytorch-crf 
scikit-learn 
tqdm
gensim

numpy==1.26.4
pandas==2.0.3

scikit-learn==1.2.2

tqdm==4.66.1